refactor: seed events domain package
