+++ pkg/agentcompose/topic_event_model.go
@@ -1,183 +1,49 @@
 package agentcompose
 
-import (
-	"crypto/sha256"
-	"encoding/hex"
-	"fmt"
-	"regexp"
-	"strings"
-	"time"
-)
+import "agent-compose/internal/agentcompose/events"
 
 const (
-	TopicEventSourceWebhook = "webhook"
-	TopicEventSourceLoader  = "loader"
-	TopicEventSourceSystem  = "system"
+	TopicEventSourceWebhook = events.TopicEventSourceWebhook
+	TopicEventSourceLoader  = events.TopicEventSourceLoader
+	TopicEventSourceSystem  = events.TopicEventSourceSystem
 
-	TopicEventDispatchPending        = "pending"
-	TopicEventDispatchPublishing     = "publishing_to_bus"
-	TopicEventDispatchPublishedToBus = "published_to_bus"
-	TopicEventDispatchNoSubscriber   = "no_subscriber"
-	TopicEventDispatchRetrying       = "retrying"
-	TopicEventDispatchDeadLetter     = "dead_letter"
+	TopicEventDispatchPending        = events.TopicEventDispatchPending
+	TopicEventDispatchPublishing     = events.TopicEventDispatchPublishing
+	TopicEventDispatchPublishedToBus = events.TopicEventDispatchPublishedToBus
+	TopicEventDispatchNoSubscriber   = events.TopicEventDispatchNoSubscriber
+	TopicEventDispatchRetrying       = events.TopicEventDispatchRetrying
+	TopicEventDispatchDeadLetter     = events.TopicEventDispatchDeadLetter
 
-	EventDeliveryStatusMatched      = "matched"
-	EventDeliveryStatusRunStarted   = "run_started"
-	EventDeliveryStatusRunSucceeded = "run_succeeded"
-	EventDeliveryStatusRunFailed    = "run_failed"
-	EventDeliveryStatusSkipped      = "skipped"
+	EventDeliveryStatusMatched      = events.EventDeliveryStatusMatched
+	EventDeliveryStatusRunStarted   = events.EventDeliveryStatusRunStarted
+	EventDeliveryStatusRunSucceeded = events.EventDeliveryStatusRunSucceeded
+	EventDeliveryStatusRunFailed    = events.EventDeliveryStatusRunFailed
+	EventDeliveryStatusSkipped      = events.EventDeliveryStatusSkipped
 )
 
-var topicEventNamePattern = regexp.MustCompile(`^[A-Za-z0-9._-]+$`)
-
-type TopicEventRecord struct {
-	ID              string    `json:"event_id"`
-	Sequence        int64     `json:"sequence"`
-	Topic           string    `json:"topic"`
-	Source          string    `json:"source"`
-	Provider        string    `json:"provider,omitempty"`
-	Intent          string    `json:"intent,omitempty"`
-	CorrelationID   string    `json:"correlation_id"`
-	IdempotencyKey  string    `json:"idempotency_key,omitempty"`
-	DeliveryID      string    `json:"delivery_id,omitempty"`
-	PayloadHash     string    `json:"payload_hash"`
-	PayloadJSON     string    `json:"payload_json"`
-	DispatchStatus  string    `json:"dispatch_status"`
-	ParentEventID   string    `json:"parent_event_id,omitempty"`
-	PublisherType   string    `json:"publisher_type,omitempty"`
-	PublisherID     string    `json:"publisher_id,omitempty"`
-	PublisherRunID  string    `json:"publisher_run_id,omitempty"`
-	ReplayOfEventID string    `json:"replay_of_event_id,omitempty"`
-	ClaimID         string    `json:"claim_id,omitempty"`
-	ClaimUntil      time.Time `json:"claim_until,omitempty"`
-	AttemptCount    int       `json:"attempt_count,omitempty"`
-	NextAttemptAt   time.Time `json:"next_attempt_at,omitempty"`
-	LastError       string    `json:"last_error,omitempty"`
-	DeadLetterAt    time.Time `json:"dead_letter_at,omitempty"`
-	CreatedAt       time.Time `json:"created_at"`
-	DispatchedAt    time.Time `json:"dispatched_at,omitempty"`
-}
-
-type TopicEventFilter struct {
-	EventID        string
-	Topic          string
-	CorrelationID  string
-	AfterSequence  int64
-	Limit          int
-	DispatchStatus string
-}
-
-type WebhookSource struct {
-	ID              string    `json:"id"`
-	Name            string    `json:"name"`
-	Enabled         bool      `json:"enabled"`
-	Provider        string    `json:"provider"`
-	TopicPrefix     string    `json:"topic_prefix"`
-	TokenHash       string    `json:"token_hash,omitempty"`
-	SignatureType   string    `json:"signature_type,omitempty"`
-	SignatureSecret string    `json:"signature_secret,omitempty"`
-	BodyLimitBytes  int64     `json:"body_limit_bytes,omitempty"`
-	CreatedAt       time.Time `json:"created_at"`
-	UpdatedAt       time.Time `json:"updated_at"`
-}
-
-type EventDelivery struct {
-	EventID   string    `json:"event_id"`
-	LoaderID  string    `json:"loader_id"`
-	TriggerID string    `json:"trigger_id"`
-	RunID     string    `json:"run_id,omitempty"`
-	Status    string    `json:"status"`
-	Error     string    `json:"error,omitempty"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-}
-
-type EventSessionLink struct {
-	EventID       string    `json:"event_id"`
-	SessionID     string    `json:"session_id"`
-	Relation      string    `json:"relation"`
-	LoaderID      string    `json:"loader_id,omitempty"`
-	RunID         string    `json:"run_id,omitempty"`
-	TriggerID     string    `json:"trigger_id,omitempty"`
-	LoaderEventID string    `json:"loader_event_id,omitempty"`
-	CreatedAt     time.Time `json:"created_at"`
-}
-
-type EventSessionTraceItem struct {
-	SessionID     string    `json:"session_id"`
-	Relation      string    `json:"relation"`
-	LoaderID      string    `json:"loader_id,omitempty"`
-	RunID         string    `json:"run_id,omitempty"`
-	TriggerID     string    `json:"trigger_id,omitempty"`
-	LoaderEventID string    `json:"loader_event_id,omitempty"`
-	EventID       string    `json:"event_id"`
-	CreatedAt     time.Time `json:"created_at"`
-}
+type TopicEventRecord = events.TopicEventRecord
+type TopicEventFilter = events.TopicEventFilter
+type WebhookSource = events.WebhookSource
+type EventDelivery = events.EventDelivery
+type EventSessionLink = events.EventSessionLink
+type EventSessionTraceItem = events.EventSessionTraceItem
 
 func validateTopicEventName(topic string) error {
-	topic = strings.TrimSpace(topic)
-	if topic == "" {
-		return fmt.Errorf("topic is required")
-	}
-	if len(topic) > 128 {
-		return fmt.Errorf("topic is too long")
-	}
-	if !topicEventNamePattern.MatchString(topic) {
-		return fmt.Errorf("topic contains invalid characters")
-	}
-	return nil
+	return events.ValidateTopicEventName(topic)
 }
 
 func normalizeTopicEventSource(source string) string {
-	switch strings.ToLower(strings.TrimSpace(source)) {
-	case TopicEventSourceWebhook:
-		return TopicEventSourceWebhook
-	case TopicEventSourceLoader:
-		return TopicEventSourceLoader
-	case TopicEventSourceSystem:
-		return TopicEventSourceSystem
-	default:
-		return ""
-	}
+	return events.NormalizeTopicEventSource(source)
 }
 
 func normalizeTopicEventDispatchStatus(status string) string {
-	switch strings.ToLower(strings.TrimSpace(status)) {
-	case "", TopicEventDispatchPending:
-		return TopicEventDispatchPending
-	case TopicEventDispatchPublishing:
-		return TopicEventDispatchPublishing
-	case TopicEventDispatchPublishedToBus:
-		return TopicEventDispatchPublishedToBus
-	case TopicEventDispatchNoSubscriber:
-		return TopicEventDispatchNoSubscriber
-	case TopicEventDispatchRetrying:
-		return TopicEventDispatchRetrying
-	case TopicEventDispatchDeadLetter:
-		return TopicEventDispatchDeadLetter
-	default:
-		return ""
-	}
+	return events.NormalizeTopicEventDispatchStatus(status)
 }
 
 func normalizeEventDeliveryStatus(status string) string {
-	switch strings.ToLower(strings.TrimSpace(status)) {
-	case EventDeliveryStatusMatched:
-		return EventDeliveryStatusMatched
-	case EventDeliveryStatusRunStarted:
-		return EventDeliveryStatusRunStarted
-	case EventDeliveryStatusRunSucceeded:
-		return EventDeliveryStatusRunSucceeded
-	case EventDeliveryStatusRunFailed:
-		return EventDeliveryStatusRunFailed
-	case EventDeliveryStatusSkipped:
-		return EventDeliveryStatusSkipped
-	default:
-		return ""
-	}
+	return events.NormalizeEventDeliveryStatus(status)
 }
 
 func topicEventPayloadSHA256(payloadJSON string) string {
-	sum := sha256.Sum256([]byte(payloadJSON))
-	return "sha256:" + hex.EncodeToString(sum[:])
+	return events.TopicEventPayloadSHA256(payloadJSON)
 }
