+++ pkg/agentcompose/topic_event_store.go
@@ -1,14 +1,13 @@
 package agentcompose
 
 import (
+	"agent-compose/internal/agentcompose/events"
 	"context"
 	"database/sql"
 	"errors"
 	"fmt"
 	"strings"
 	"time"
-
-	"github.com/google/uuid"
 )
 
 func (s *ConfigStore) ensureEventSchema(ctx context.Context) error {
@@ -116,72 +115,7 @@ func (s *ConfigStore) ensureEventSchema(ctx context.Context) error {
 }
 
 func normalizeTopicEventRecord(item TopicEventRecord, assignID bool) (TopicEventRecord, error) {
-	item.ID = strings.TrimSpace(item.ID)
-	if assignID && item.ID == "" {
-		item.ID = "evt_" + uuid.NewString()
-	}
-	if item.ID == "" {
-		return TopicEventRecord{}, fmt.Errorf("event id is required")
-	}
-	item.Topic = strings.TrimSpace(item.Topic)
-	if err := validateTopicEventName(item.Topic); err != nil {
-		return TopicEventRecord{}, err
-	}
-	item.Source = normalizeTopicEventSource(item.Source)
-	if item.Source == "" {
-		return TopicEventRecord{}, fmt.Errorf("event source is required")
-	}
-	item.DispatchStatus = normalizeTopicEventDispatchStatus(item.DispatchStatus)
-	if item.DispatchStatus == "" {
-		return TopicEventRecord{}, fmt.Errorf("event dispatch status is invalid")
-	}
-	item.Provider = strings.TrimSpace(item.Provider)
-	item.Intent = strings.TrimSpace(item.Intent)
-	item.CorrelationID = strings.TrimSpace(item.CorrelationID)
-	if item.CorrelationID == "" {
-		item.CorrelationID = item.ID
-	}
-	item.IdempotencyKey = strings.TrimSpace(item.IdempotencyKey)
-	item.DeliveryID = strings.TrimSpace(item.DeliveryID)
-	item.PayloadJSON = strings.TrimSpace(item.PayloadJSON)
-	if item.PayloadJSON == "" {
-		item.PayloadJSON = "{}"
-	}
-	if _, err := normalizeJSONDocument(item.PayloadJSON); err != nil {
-		return TopicEventRecord{}, err
-	}
-	item.PayloadHash = strings.TrimSpace(item.PayloadHash)
-	if item.PayloadHash == "" {
-		item.PayloadHash = topicEventPayloadSHA256(item.PayloadJSON)
-	}
-	item.ParentEventID = strings.TrimSpace(item.ParentEventID)
-	item.PublisherType = strings.TrimSpace(item.PublisherType)
-	item.PublisherID = strings.TrimSpace(item.PublisherID)
-	item.PublisherRunID = strings.TrimSpace(item.PublisherRunID)
-	item.ReplayOfEventID = strings.TrimSpace(item.ReplayOfEventID)
-	item.ClaimID = strings.TrimSpace(item.ClaimID)
-	if !item.ClaimUntil.IsZero() {
-		item.ClaimUntil = item.ClaimUntil.UTC()
-	}
-	if item.AttemptCount < 0 {
-		item.AttemptCount = 0
-	}
-	if !item.NextAttemptAt.IsZero() {
-		item.NextAttemptAt = item.NextAttemptAt.UTC()
-	}
-	item.LastError = strings.TrimSpace(item.LastError)
-	if !item.DeadLetterAt.IsZero() {
-		item.DeadLetterAt = item.DeadLetterAt.UTC()
-	}
-	if item.CreatedAt.IsZero() {
-		item.CreatedAt = time.Now().UTC()
-	} else {
-		item.CreatedAt = item.CreatedAt.UTC()
-	}
-	if !item.DispatchedAt.IsZero() {
-		item.DispatchedAt = item.DispatchedAt.UTC()
-	}
-	return item, nil
+	return events.NormalizeTopicEventRecord(item, assignID)
 }
 
 func (s *ConfigStore) CreateEvent(ctx context.Context, item TopicEventRecord) (TopicEventRecord, error) {
@@ -799,15 +733,7 @@ func (s *ConfigStore) ListEnabledWebhookSourcesForTopic(ctx context.Context, top
 }
 
 func webhookSourceTopicMatches(topic, topicPrefix string) bool {
-	topic = strings.TrimSpace(topic)
-	topicPrefix = strings.TrimSpace(topicPrefix)
-	if topic == "" || topicPrefix == "" {
-		return false
-	}
-	if strings.HasPrefix(topic, topicPrefix) {
-		return true
-	}
-	return strings.HasSuffix(topicPrefix, ".") && topic == strings.TrimSuffix(topicPrefix, ".")
+	return events.WebhookSourceTopicMatches(topic, topicPrefix)
 }
 
 func (s *ConfigStore) ListWebhookSources(ctx context.Context) ([]WebhookSource, error) {
